Removed addAppointment method
Removed the addAppointment method from the interface and class for Patient, since it was discussed as not wanted anymore, similar to the Doctor's one. I've also made a list of things we want changed from the original plan (which is what I refer to the image as) so that we can keep track and I've implemented the changes I've listed already.